+++ docs/release-notes.md
@@ -5,6 +5,22 @@ hide:
 
 # Release Notes
 
+## Unreleased
+
+### Added
+
+- `default_timezone`, `force_timezone`, `remove_timezone` for DateTimeField
+- `default_timezone`, `force_timezone` for DateField.
+-  Add attribute `inject_default_on_partial_update`.
+
+### Changed
+
+- `get_default_values` has now an extra keyword argument `is_update`
+
+### Fixed
+
+- `auto_now` and `auto_now_add` now also work for date fields
+
 ## 0.12.0
 
 ### Added